# docs/swagger/swagger.yaml
basePath: /v1
consumes:
    - application/json
definitions:
    Group:
        properties:
            created_at:
                description: The timestamp of when the group was created.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: CreatedAt
            deleted_at:
                description: The timestamp of when the group was deleted.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: DeletedAt
            id:
                description: The ID of the group.
                example: 1
                format: uint64
                type: integer
                x-go-name: ID
            name:
                description: The name of the group.
                example: '"My Group"'
                type: string
                x-go-name: Name
            updated_at:
                description: The timestamp of when the group was last updated.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: UpdatedAt
            uuid:
                description: The UUID of the group.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: UUID
        required:
            - id
            - uuid
            - name
        title: Group represents a group in the system.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/response
    GroupRequest:
        properties:
            created_at:
                description: The timestamp of when the group was created.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: CreatedAt
            deleted_at:
                description: The timestamp of when the group was deleted.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: DeletedAt
            id:
                description: The ID of the group.
                example: 1
                format: uint64
                type: integer
                x-go-name: ID
            name:
                description: The name of the group.
                example: '"My Group"'
                type: string
                x-go-name: Name
            updated_at:
                description: The timestamp of when the group was last updated.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: UpdatedAt
            uuid:
                description: The UUID of the group.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: UUID
        required:
            - name
        title: GroupRequest represents the request body for group-related operations.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/request
    GroupResponse:
        properties:
            code:
                description: The response code.
                example: '"SUCCESS"'
                type: string
                x-go-name: Code
            groups:
                description: The list of groups.
                items:
                    $ref: '#/definitions/Group'
                type: array
                x-go-name: Groups
            message:
                description: The response message.
                example: '"Groups retrieved successfully"'
                type: string
                x-go-name: Message
        required:
            - code
            - message
            - groups
        title: GroupResponse represents the response body for group-related operations.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/response
...
host: localhost:8080
info:
    description: '# This is a sample Locky API'
    title: Locky API.
    version: 1.0.0
paths:
...
    /private/groups:
        get:
            description: Get a list of all groups in the system.
            operationId: getGroupsPrivate
            responses:
                "200":
                    description: A list of groups.
                    schema:
                        $ref: '#/definitions/GroupResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/GroupResponse'
            summary: Get a list of groups.
            tags:
                - groups
        post:
            description: Create a new group with the provided information.
            operationId: createGroupPrivate
            parameters:
                - description: The group to create.
                  in: body
                  name: group
                  required: true
                  schema:
                    $ref: '#/definitions/GroupRequest'
            responses:
                "200":
                    description: The created group.
                    schema:
                        $ref: '#/definitions/GroupResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/GroupResponse'
            summary: Create a new group.
            tags:
                - groups
    /private/groups/{id}:
        delete:
            description: Delete a group by ID.
            operationId: deleteGroupPrivate
            parameters:
                - description: The ID of the group to delete.
                  in: path
                  name: id
                  required: true
                  type: integer
            responses:
                "200":
                    description: The deleted group.
                    schema:
                        $ref: '#/definitions/GroupResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/GroupResponse'
            summary: Delete a group.
            tags:
                - groups
        put:
            description: Update a group with the provided information.
            operationId: updateGroupPrivate
            parameters:
                - description: The ID of the group to update.
                  in: path
                  name: id
                  required: true
                  type: integer
                - description: The group to update.
                  in: body
                  name: group
                  required: true
                  schema:
                    $ref: '#/definitions/GroupRequest'
            responses:
                "200":
                    description: The updated group.
                    schema:
                        $ref: '#/definitions/GroupResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/GroupResponse'
            summary: Update a group.
            tags:
                - groups
...
    /public/groups:
        get:
            description: Get a list of all groups in the system.
            operationId: getGroups
            responses:
                "200":
                    description: A list of groups.
                    schema:
                        $ref: '#/definitions/GroupResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/GroupResponse'
            summary: Get a list of groups.
            tags:
                - groups
...
produces:
    - application/json
schemes:
    - http
swagger: "2.0"